feat: move protobuf from "playback" to "lib"
