+++ tui/src/ui/mod.rs
@@ -31,8 +31,8 @@ use model::{Model, TermusicLayout};
 use playback::Playback;
 use std::time::Duration;
 use sysinfo::System;
+use termusiclib::player::music_player_client::MusicPlayerClient;
 pub use termusiclib::types::*;
-use termusicplayback::player::music_player_client::MusicPlayerClient;
 use termusicplayback::{PlayerCmd, PlayerProgress, Status};
 use tokio::sync::mpsc::{self, UnboundedReceiver};
 use tonic::transport::Channel;
